+++ test/function.ts
@@ -14,7 +14,8 @@ import {
   absurd,
   flow,
   tupled,
-  untupled
+  untupled,
+  pipe
 } from '../src/function'
 
 const f = (n: number) => n + 1
@@ -99,5 +100,23 @@ describe('function', () => {
     const u2 = untupled(f2)
     assert.deepStrictEqual(u1(1), 2)
     assert.deepStrictEqual(u2(1, 2), 3)
+  })
+
+  it('pipe', () => {
+    const f = (n: number) => n + 1
+    const g = (n: number) => n * 2
+    assert.deepStrictEqual(pipe(2), 2)
+    assert.deepStrictEqual(pipe(2, f), 3)
+    assert.deepStrictEqual(pipe(2, f, g), 6)
+    assert.deepStrictEqual(pipe(2, f, g, f), 7)
+    assert.deepStrictEqual(pipe(2, f, g, f, g), 14)
+    assert.deepStrictEqual(pipe(2, f, g, f, g, f), 15)
+    assert.deepStrictEqual(pipe(2, f, g, f, g, f, g), 30)
+    assert.deepStrictEqual(pipe(2, f, g, f, g, f, g, f), 31)
+    assert.deepStrictEqual(pipe(2, f, g, f, g, f, g, f, g), 62)
+    assert.deepStrictEqual(pipe(2, f, g, f, g, f, g, f, g, f), 63)
+    assert.deepStrictEqual(pipe(2, f, g, f, g, f, g, f, g, f), 63)
+    // this is just to satisfy noImplicitReturns and 100% coverage
+    assert.deepStrictEqual((pipe as any)(...[2, f, g, f, g, f, g, f, g, f, g]), undefined)
   })
 })