+++ CHANGELOG.md
@@ -18,6 +18,7 @@ high state of flux, you're at risk of it changing without notice.
 
 - **Polish**
   - change `ReadonlyNonEmptyArray` to get better type inference (@gcanti)
+  - move `pipe` to `function` module (@gcanti)
 
 # 2.6.2